Add search button on map and remove auto search.
